+++ lib/src/obsoleted/rules/avoid_unused_parameters.dart
@@ -26,6 +26,7 @@ class AvoidUnusedParameters extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
         );
 
   @override