+++ lib/src/obsoleted/rules/component_annotation_arguments_ordering.dart
@@ -25,6 +25,7 @@ class ComponentAnnotationArgumentsOrderingRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.style),
+          excludes: readExcludes(config),
         );
 
   @override