+++ lib/src/obsoleted/rules/member_ordering.dart
@@ -29,6 +29,7 @@ class MemberOrderingRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.style),
+          excludes: readExcludes(config),
         );
 
   @override