+++ lib/src/obsoleted/rules/newline_before_return.dart
@@ -23,6 +23,7 @@ class NewlineBeforeReturnRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.style),
+          excludes: readExcludes(config),
         );
 
   @override