+++ lib/src/obsoleted/rules/no_empty_block.dart
@@ -22,6 +22,7 @@ class NoEmptyBlockRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.style),
+          excludes: readExcludes(config),
         );
 
   @override