+++ lib/src/obsoleted/rules/no_equal_arguments.dart
@@ -19,6 +19,7 @@ class NoEqualArguments extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
         );
 
   @override