+++ lib/src/obsoleted/rules/no_magic_number_rule.dart
@@ -23,6 +23,7 @@ class NoMagicNumberRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
         );
 
   @override