+++ lib/src/obsoleted/rules/prefer_intl_name.dart
@@ -26,6 +26,7 @@ class PreferIntlNameRule extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
         );
 
   @override