+++ lib/src/obsoleted/rules/prefer_trailing_comma.dart
@@ -27,6 +27,7 @@ class PreferTrailingComma extends ObsoleteRule {
           id: ruleId,
           documentationUrl: Uri.parse(_documentationUrl),
           severity: readSeverity(config, Severity.warning),
+          excludes: readExcludes(config),
         );
 
   @override